Update README.md
Backorder is an order which can’t be fulfilled at the given time due to lack of supply or the product is currently out of stock or not in inventory but can guarantee delivery of the goods or service requested by a certain date in the future because the production of goods or replenishment of inventory is underway. Unlike in the situation of Out-of-stock where the delivery date of the goods can’t be promised, in the Backorder scenario, the customers are allowed to shop for the products and order. Simply put Backorder can be thought of as an order with a delayed delivery date